refactoring icon generation
https://coveord.atlassian.net/browse/KIT-812
